Added an function in shell! Try to write mudstorm on shell :D
New version of shell(0.5) & OS(0.6.2)

+++ kernel/include/delay.h
@@ -21,6 +21,6 @@
 extern void DelayS(int s);
 extern void DelayMs(int ms);
 
-extern int cpu_speed();
+extern unsigned long cpu_speed();
 
 #endif
